refactored model and book model
